requirements.txt
boto3==1.3.0
botocore==1.4.24
docutils==0.12
futures==3.0.5
Jinja2==2.8
jmespath==0.9.0
lambda-uploader==1.0.0
MarkupSafe==0.23
python-dateutil==2.5.3
PyYAML==3.11
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.11
six==1.10.0
virtualenv==15.0.2
yoke==0.0.1
